+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # vNext
 - Fix for compile-time crash when mapped type has no `aliasTypeArguments`
 - Fix for properties which use a string literal name
+- Add missing support for class expressions
 
 # v0.4.14
 - Support for metadata on properties with computed names (including symbols)